\name{lbpc}
\alias{lbpc}
\title{ List the Functions of the BioPhysConnectoR Package }
\description{
  A shortcut for ls("package:BioPhysConnectoR") to get an overview of the implemented functions.
}
\usage{
lbpc()
}
\value{
 No return value. Prints the function names.
}

\author{ Franziska Hoffgaard }
\keyword{ utilities }
